Rename routeRank to camerCase
