+++ packages/agents-hosting/src/app/extensions.ts
@@ -1,7 +1,7 @@
 import { TurnContext } from '../turnContext'
 import { AgentApplication } from './agentApplication'
 import { RouteHandler } from './routeHandler'
-import { RouteRank } from './RouteRank'
+import { RouteRank } from './routeRank'
 import { RouteSelector } from './routeSelector'
 import { TurnState } from './turnState'
